refactor: changed destination of type now types are in the original destination where type is defined for easier changes to code/ i also refactored features folder structure
