+++ src/features/adminPanel/CompanyManagment/CompanyManagment.tsx
@@ -2,7 +2,7 @@ import { Label } from "@/components/ui/label"
 import { Input } from "@/components/ui/input"
 import { Button } from "@/components/ui/button"
 import React, { useState } from "react"
-import { CompanyManagmentOverview } from "@/features/adminPanel-CompanyManagment/CompanyManagmentOverview"
+import { CompanyManagmentOverview } from "@/features/adminPanel/CompanyManagment/CompanyManagmentOverview"
 import { CompanyManagmentData } from "@/data/adminManagment/CompanyManagmentData"
 import { TFormField, handleChangeInput } from "@/lib/utils/HandlingChangeInput"
 import { saveCompanyEdited } from "@/lib/utils/saveEditedCompany"
