+++ src/lib/utils/LoginAuthentication.ts
@@ -1,4 +1,4 @@
-import { TLoginCredentials } from '@/data/dataTypes/user-creds-types-d';
+import { TLoginCredentials } from '@/features/authentication/Login';
 import axios from 'axios';
 
 type AuthSetter = (token: string, email: string, userName: string) => void;
