+++ src/pages/privatePageWrappers/UserProtectedWrapper.tsx
@@ -11,6 +11,7 @@ const UserProtectedWrapper = () => {
 
   useEffect(() => {
     const handleAuthentication = () => {
+      //! if localStorageToken is set with starting "!"- means it checks for token if you remove this you can use this application without logging in
       if (localStorageToken) {
         navigate("/login");
       }
@@ -23,9 +24,7 @@ const UserProtectedWrapper = () => {
       <Navbar setIsOpen={setIsOpen} />
       <div className="flex w-full">
         <Sidebar isOpen={isOpen} setIsOpen={setIsOpen} />
-        <div className="flex w-full">
-          <Outlet />
-        </div>
+        <Outlet />
       </div>
     </div>
   );
